+++ lib/src/helpers/helpers.dart
@@ -1,3 +1,4 @@
 export 'src/isolate.dart';
 export 'src/result.dart';
 export 'src/logger.dart';
+export 'src/page_animation.dart';