+++ CHANGELOG.md
@@ -1,4 +1,9 @@
+## 0.0.1-dev.3
+- Added the PageAnimation from custom_widgets_toolkit
+- Removed specific state_mgmt modules
+
 ## 0.0.1-dev.2
 - Fixed exporting of internal package indiscriminately
+
 ## 0.0.1-dev.1
 - First publish. Abstracted away from the package kickin